fixed type of in container names, removed manta and cnvkit from workflow

+++ lib/wf_agg_cnv.nf
@@ -43,6 +43,28 @@ process savvy_to_bed{
     """
 }
 
+process cnvkit_to_bed{
+    tag {idPatient + "-" + idSample}
+    label 'container_llab'
+
+    publishDir "${params.outdir}/VariantCalling/${idSample}/ReCalToMarkedRaw/", mode: params.publish_dir_mode
+
+    input:
+    tuple idPatient, idSample, file(cnr)
+    file(exon_file)
+
+    output:
+    tuple val('cnvkit'), idPatient, idSample, file("${idSample}.bed")
+
+    script:
+    """
+        # strip the first line # the next line is hard coded and needs to be fix before production as does the input parameter
+        sed '1 d' /scratch/Shares/CHCO/workspace/cna_positive_wes/results/cnvkit/VariantCalling/${idSample}/CNVKit/${idSample}.cnr | bedtools intersect -wb -b stdin -a $exon_file > cnvtemp.tsv
+        # sed '1 d' $cnr | bedtools intersect -wb -b stdin -a $exon_file > cnvtemp.tsv
+        python cnv-kit_to_bed.py $idSample ${idSample}.bed cnvtemp.tsv
+    """
+}
+
 process combine_callers{
     tag {idPatient + "-" + idSample}
     label 'container_llab'
@@ -50,7 +72,8 @@ process combine_callers{
     publishDir "${params.outdir}/VariantCalling/${idSample}/AllCNVCallers/", mode: params.publish_dir_mode
 
     input:
-    tuple caller, idPatient, idSample, file(manta_bed)
+    //tuple caller, idPatient, idSample, file(manta_bed)
+    tuple val(idSample)
     file(savvy_beds)
     // add input for CNVKIT eventually
 
@@ -59,19 +82,25 @@ process combine_callers{
 
     script:
     """
-    cat $manta_bed >> temp.bed
-
     FILE="${idSample}.coverageBinner_savvy.bed"
     if test -f "\$FILE"; then
         cat ${idSample}.coverageBinner_savvy.bed >> temp.bed
     fi 
 
-    #cnvkit_file=\$(find )
+    if test -f "temp.bed"; then
+        #cnvkit_file=\$(find )
 
-    grep -v BND temp.bed | awk '(\$2 <= \$3)' >  filtered_temp.bed
-    cat filtered_temp.bed | sort -k1,1V -k2,2n -k3,3n > tripple_sorted.bed
-    bedtools cluster -i tripple_sorted.bed > clustered_test.bed
-    agg_cluster.py clustered_test.bed > ${idSample}.multi_caller.bed
+        grep -v BND temp.bed | awk '(\$2 <= \$3)' >  filtered_temp.bed
+        echo "2"
+        cat filtered_temp.bed | sort -k1,1V -k2,2n -k3,3n > tripple_sorted.bed
+        echo "3"
+        bedtools cluster -i tripple_sorted.bed > clustered_test.bed
+        echo "4"
+        agg_cluster.py clustered_test.bed > ${idSample}.multi_caller.bed
+        echo "5"
+    else
+        touch ${idSample}.multi_caller.bed
+    fi
     """
 }
